requirements.txt
mcp>=1.0.0
PyGithub>=2.1.1
python-dotenv>=1.0.0
pytest>=7.4.0
requests>=2.31.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
